Cast port variable to uint16_t in htons call to prevent compiler conversion warning

+++ src/main.cpp
@@ -129,7 +129,7 @@ int main(int argc, char* argv[]) {
     sockaddr_in address{};
     address.sin_family = AF_INET;
     address.sin_addr.s_addr = INADDR_ANY;
-    address.sin_port = htons(port);
+    address.sin_port = htons(static_cast<uint16_t>(port));
 
     if (bind(server_fd, (struct sockaddr*)&address, sizeof(address)) < 0) {
         std::cerr << "Bind failed\n";
